Bugfix: Float decimal point for multilanguage setups
In multilanguage setups and when using PHP setlocale() the decimal point changes from a dot (.) to a comma (,) in some languages (e.g. German). This means that "setAspectRatio" no longer works. This is now fixed and also the faster variant "(float)" was used for the type conversion. Additionally, round() was used to limit the number of decimal places to 5. Depending on the ratio, the decimal places can become unnecessarily long.